+++ packages/cli/CHANGELOG.md
@@ -29,6 +29,8 @@
 #### 🐞 Fixes
 
 - Fixed an issue where "installing yarn" would constantly show.
+- Fixed an issue on Windows where `package.json` and `tsconfig.json` would be modified and linefeeds
+  would change.
 
 ## 0.20.3